// src/app/client/model/productVariation.ts
/**
 * my api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VariationId } from './variationId';


export class ProductVariation {
    withoutCount?: boolean;
    id?: string;
    title?: string;
    isArchived?: boolean;
    canBeOrdered?: boolean;
    inventoryCount?: number;
    price?: number;
    variationIds?: Array<VariationId>;
}